pyproject.toml
[tool.poetry]
name = "toucans-hub"
version = "0.1.0"
description = "Prompt-hub for the toucans PyPI package."
authors = ["Kasper Junge"]
readme = "README.md"
packages = [{include = "toucans_hub"}]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = "^0.104.1"
sqlmodel = "^0.0.12"
alembic = "^1.12.1"
python-dotenv = "^1.0.0"
uvicorn = "^0.24.0.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
